tests: Improve failure reporting by using then-catch.
Previously, there was a huge error in the test framework that even
chai threw an error and reected the promise, we didn't catch it and
fail the test. Instead, it just logged an unhandled Promise rejection
warning in the console. Thus, any failing tests of promises could be
detected only by looking at the tests when running them. This PR
fixes this for all the test files.